Move subchannel_call_holder
