+++ build.yaml
@@ -262,7 +262,6 @@ filegroups:
   - src/core/lib/channel/context.h
   - src/core/lib/channel/http_client_filter.h
   - src/core/lib/channel/http_server_filter.h
-  - src/core/lib/channel/subchannel_call_holder.h
   - src/core/lib/compression/algorithm_metadata.h
   - src/core/lib/compression/message_compress.h
   - src/core/lib/debug/trace.h
@@ -336,7 +335,6 @@ filegroups:
   - src/core/lib/channel/connected_channel.c
   - src/core/lib/channel/http_client_filter.c
   - src/core/lib/channel/http_server_filter.c
-  - src/core/lib/channel/subchannel_call_holder.c
   - src/core/lib/compression/compression_algorithm.c
   - src/core/lib/compression/message_compress.c
   - src/core/lib/debug/trace.c
@@ -426,6 +424,7 @@ filegroups:
   - src/core/ext/client_config/resolver_factory.h
   - src/core/ext/client_config/resolver_registry.h
   - src/core/ext/client_config/subchannel.h
+  - src/core/ext/client_config/subchannel_call_holder.h
   - src/core/ext/client_config/subchannel_index.h
   - src/core/ext/client_config/uri_parser.h
   src:
@@ -443,6 +442,7 @@ filegroups:
   - src/core/ext/client_config/resolver_factory.c
   - src/core/ext/client_config/resolver_registry.c
   - src/core/ext/client_config/subchannel.c
+  - src/core/ext/client_config/subchannel_call_holder.c
   - src/core/ext/client_config/subchannel_index.c
   - src/core/ext/client_config/uri_parser.c
 - name: grpc_codegen
